first run at getting the new styles integrated
