+++ gatsby-config.js
@@ -32,7 +32,6 @@ module.exports = {
     `gatsby-plugin-stylus`,
     `gatsby-plugin-react-helmet`,
     `gatsby-plugin-flow`,
-    `gatsby-plugin-layout`,
     `gatsby-plugin-sass`
   ],
 }
