+++ src/css/_base.scss
@@ -1,0 +1,385 @@
+/**
+ * Global styles, the starting point. No classes or ids.
+ * contains our resets, link styling, default font sizes, etc
+*/
+ *,
+ *:before,
+ *:after {
+   box-sizing: border-box;
+ }
+ /**
+  * universal baseline
+  * 1 font-smoothing makes our text sharp
+  * 2 wont work on firefox unfortunately
+  * 3 this setting can speed up font rendering significantly, esp on mobile
+  * 4 remove click delay on microsoft devices
+  */
+ /* http://meyerweb.com/eric/tools/css/reset/
+    v2.0 | 20110126
+    License: none (public domain)
+ */
+ html,
+ body,
+ div,
+ span,
+ applet,
+ object,
+ iframe,
+ h1,
+ h2,
+ h3,
+ h4,
+ h5,
+ h6,
+ p,
+ blockquote,
+ pre,
+ a,
+ abbr,
+ acronym,
+ address,
+ big,
+ cite,
+ code,
+ del,
+ dfn,
+ em,
+ img,
+ ins,
+ kbd,
+ q,
+ s,
+ samp,
+ small,
+ strike,
+ strong,
+ sub,
+ sup,
+ tt,
+ var,
+ b,
+ u,
+ i,
+ center,
+ dl,
+ dt,
+ dd,
+ ol,
+ ul,
+ li,
+ fieldset,
+ form,
+ label,
+ legend,
+ table,
+ caption,
+ tbody,
+ tfoot,
+ thead,
+ tr,
+ th,
+ td,
+ article,
+ aside,
+ canvas,
+ details,
+ embed,
+ figure,
+ figcaption,
+ footer,
+ header,
+ hgroup,
+ menu,
+ nav,
+ output,
+ ruby,
+ section,
+ summary,
+ time,
+ mark,
+ audio,
+ video {
+   margin: 0;
+   padding: 0;
+   border: 0;
+   font-size: 100%;
+   font: inherit;
+   vertical-align: baseline;
+ }
+ /* HTML5 display-role reset for older browsers */
+ article,
+ aside,
+ details,
+ figcaption,
+ figure,
+ footer,
+ header,
+ hgroup,
+ menu,
+ nav,
+ section {
+   display: block;
+ }
+ 
+ body {
+   line-height: 1;
+ }
+ 
+ html {
+   -moz-osx-font-smoothing: grayscale;
+   -ms-text-size-adjust: 100%;
+   -webkit-font-smoothing: antialiased;
+   -webkit-osx-font-smoothing: grayscale;
+   -webkit-overflow-scrolling: touch;
+   -webkit-text-size-adjust: 100%;
+   background-color: $clrWhite;
+   color: $clrBase;
+   font-family: $sans;
+   font-size: $pSize;
+   font-weight: 300;
+   overflow-x: hidden;
+   scroll-behavior: smooth;
+   text-rendering: optimizeSpeed;
+   touch-action: manipulation;
+   overflow-y: scroll;
+ }
+ 
+ @media (min-width: $bpTab) {
+   html {
+     text-rendering: optimizeLegibility;
+   }
+ }
+ 
+ a {
+   color: $clrPrimary;
+   cursor: pointer;
+   text-decoration: none;
+ }
+ 
+ p {
+   color: $clrBase;
+   font-size: $pSize;
+   line-height: $pLH;
+   margin-bottom: 20px;
+ }
+ 
+ h1,
+ h2,
+ h3,
+ h4,
+ h5,
+ h6 {
+   font-family: $serif;
+   font-weight: 700;
+ }
+ 
+ ul,
+ ol,
+ li {
+   list-style: none;
+ }
+ 
+ li.bullet {
+   display: table-row;
+ }
+ 
+ li.bullet:before {
+   content: '\2022';
+   display: table-cell;
+   padding-right: 8px;
+ }
+ 
+ li.counter {
+   counter-increment: table-ol;
+ }
+ 
+ li.counter:before {
+   content: counter(table-ol) '.';
+   display: table-cell;
+   padding-right: 8px;
+   text-align: right;
+ }
+ 
+ select,
+ input[type='submit'],
+ button,
+ svg {
+   cursor: pointer;
+ }
+ 
+ button,
+ input,
+ select,
+ option,
+ textarea {
+   letter-spacing: 0.025em;
+ }
+ 
+ button,
+ select,
+ option {
+   -webkit-appearance: caret;
+   appearance: none;
+ }
+ 
+ .btn,
+ button {
+   border: 0;
+   border-radius: 5px;
+   color: $clrGrayDark;
+   background: transparent;
+   cursor: pointer;
+   display: inline-block;
+   font-family: $sans;
+   font-size: $pSize;
+   outline: 0;
+   padding: 15px 20px;
+ }
+ 
+ .btn:focus,
+ button:focus,
+ .btn:active,
+ button:active {
+   box-shadow: 0 0 3px $clrPrimary, 0 0 7px $clrPrimary;
+ }
+ 
+ .btn .btn,
+ button .btn {
+   letter-spacing: 1px;
+ }
+ 
+ .btn-icon-right {
+   transition: all 0.2s linear;
+ }
+ 
+ .btn-icon-right i {
+   transition: all 0.2s linear;
+ }
+ 
+ .btn-icon-right:hover i {
+   transform: translateX(4px);
+ }
+ 
+ select,
+ .select {
+   appearance: none;
+   background: $clrWhite;
+   border-radius: 0;
+   color: $clrGrayDark;
+   display: block;
+   font-family: inherit;
+   font-size: $h4Size;
+   padding: 7px;
+   width: 100%;
+ }
+ 
+ select {
+ /* for Firefox */
+   -moz-appearance: none;
+ /* for Chrome */
+   -webkit-appearance: none;
+ }
+ /* For IE10 */
+ select::-ms-expand {
+   display: none;
+ }
+ 
+ img,
+ iframe,
+ video {
+   display: block;
+   height: auto;
+   max-width: 100%;
+ }
+ 
+ strong {
+   font-weight: 700;
+ }
+ 
+ pre,
+ code {
+   background: $clrGrayLightest;
+   color: $clrBase;
+   font-family: $mono;
+   font-size: $smallSize;
+ }
+ 
+ .reverse-pre pre,
+ .reverse-pre code {
+   background: $clrGrayDark;
+   color: $clrGrayLightest;
+ }
+ 
+ code {
+   border-radius: 3px;
+   display: inline-block;
+   padding: 0 5px;
+   white-space: pre-wrap;
+ }
+ 
+ li.usage-api-li {
+   list-style: disc outside none;
+ }
+ 
+ .body-doc-container {
+   display: flex;
+ }
+
+ .font-size-1 {
+  font-size: $h2Size;
+  line-height: $h2LH;
+  @media (min-width: $bpTab) {
+    font-size: $h1Size;
+    line-height: $h1LH;
+  }
+}
+
+.font-size-2 {
+  font-size: $h3Size;
+  line-height: $h3LH;
+  @media (min-width: $bpTab) {
+    font-size: $h2Size;
+    line-height: $h2LH;
+  }
+}
+
+.font-size-3 {
+  font-size: $h4Size;
+  line-height: $h4LH;
+  @media (min-width: $bpTab) {
+    font-size: $h3Size;
+    line-height: $h3LH;
+  }
+}
+
+.font-size-4 {
+  font-size: $ledeSize;
+  line-height: $ledeLH;
+}
+
+.font-size-5 {
+  font-size: $pSize;
+  line-height: $pLH;
+}
+
+.small {
+  font-size: $smallSize;
+  line-height: $smallLH;
+}
+
+.smallest {
+  font-size: $smallestSize;
+  line-height: $smallestLH;
+}
+
+ 
+ @media (max-width: $tablet) {
+   .body-container {
+     padding-top: 128px;
+   }
+ 
+   .body-doc-container {
+     display: flex;
+     flex-direction: column;
+   }
+ }
