+++ src/css/_helpers.scss
@@ -1,0 +1,368 @@
+.sans {
+	font-family: $sans;
+}
+
+.serif {
+	font-family: $serif;
+}
+
+.mono {
+	font-family: $mono;
+}
+
+.small-caps {
+	font-variant: small-caps;
+}
+
+.no-b {
+	border: 0 !important;
+}
+
+.body-bg-offwhite {
+	background-color: #f3f4f4;
+}
+
+.b-1 {
+	border: 1px solid $clrGrayLighter;
+}
+
+.b-t-1 {
+	border-top: 1px solid $clrGrayLighter;
+}
+
+.b-r-1 {
+	border-right: 1px solid $clrGrayLighter;
+}
+
+.b-b-1 {
+	border-bottom: 1px solid $clrGrayLighter;
+}
+
+.b-l-1 {
+	border-left: 1px solid $clrGrayLighter;
+}
+
+.b-t-light-1 {
+	border-top: 1px solid $clrGrayLighter2;
+}
+
+.b-r-light-1 {
+	border-right: 1px solid $clrGrayLighter2;
+}
+
+.b-b-light-1 {
+	border-bottom: 1px solid $clrGrayLighter2;
+}
+
+.b-l-light-1 {
+	border-left: 1px solid $clrGrayLighter2;
+}
+
+.b-2 {
+	border: 2px solid $clrGrayLighter;
+}
+
+.b-t-2 {
+	border-top: 2px solid $clrGrayLighter;
+}
+
+.b-r-2 {
+	border-right: 2px solid $clrGrayLighter;
+}
+
+.b-b-2 {
+	border-bottom: 2px solid $clrGrayLighter;
+}
+
+.b-l-2 {
+	border-left: 2px solid $clrGrayLighter;
+}
+
+.b-t-light-2 {
+	border-top: 2px solid $clrGrayLighter2;
+}
+
+.b-r-light-2 {
+	border-right: 2px solid $clrGrayLighter2;
+}
+
+.b-b-light-2 {
+	border-bottom: 2px solid $clrGrayLighter2;
+}
+
+.b-l-light-1 {
+	border-left: 2px solid $clrGrayLighter2;
+}
+
+.center {
+	display: block;
+	margin-left: auto;
+	margin-right: auto;
+}
+
+.cursor-pointer {
+	cursor: pointer;
+}
+
+.no-marg {
+	margin: 0 !important;
+}
+
+.txt-l {
+	text-align: left;
+}
+
+.txt-c {
+	text-align: center;
+}
+
+.txt-r {
+	text-align: right;
+}
+
+.txt-overflow-ellipsis {
+	text-overflow: ellipsis;
+}
+
+.italic {
+	font-style: italic !important;
+}
+
+.float-l {
+	float: left;
+}
+
+.float-r {
+	float: right;
+}
+
+.underline {
+	text-decoration: underline;
+}
+
+.no-underline {
+	text-decoration: none;
+}
+
+.tab {
+	border-bottom-left-radius: 0;
+	border-bottom-right-radius: 0;
+}
+/**
+ * @desc color helperz
+ * @returns color classes for every color in the hash
+ */
+
+ @each $key, $value in $uswds-colors {
+  .bg-#{$key} {
+    background-color: $value;
+  }
+  
+  .hvr-bg-#{$key}.active,
+  .hvr-bg-#{$key}:hover,
+  .hvr-bg-#{$key}:active {
+    background-color: $value;
+  }
+  
+  .clr-#{$key} {
+    color: $value;
+  }
+  
+  .hvr-clr-#{$key}.active,
+  .hvr-clr-#{$key}:hover,
+  .hvr-clr-#{$key}:active {
+    color: $value;
+  }
+}
+
+.row,
+.block {
+	display: block;
+}
+
+.inline-block {
+	display: inline-block;
+}
+
+.absolute {
+	position: absolute;
+}
+
+.fixed {
+	position: fixed;
+}
+
+.relative {
+	position: relative;
+}
+
+.top {
+	top: 0;
+}
+
+.right {
+	right: 0;
+}
+
+.bottom {
+	bottom: 0;
+}
+
+.left {
+	left: 0;
+}
+
+.overflow {
+	overflow: hidden;
+}
+
+.weight-300 {
+	font-weight: 300;
+}
+
+.weight-400 {
+	font-weight: 400;
+}
+
+.weight-600 {
+	font-weight: 600;
+}
+
+.weight-700 {
+	font-weight: 700;
+}
+
+.overflow-hidden {
+	overflow: hidden;
+}
+
+.overflow-scroll {
+	overflow-y: scroll;
+}
+
+.always-show-scroll::-webkit-scrollbar {
+	appearance: none;
+	width: 7px;
+}
+
+.always-show-scroll::-webkit-scrollbar-track {
+	background-color: transparent;
+}
+
+.always-show-scroll::-webkit-scrollbar-thumb {
+	background-color: $clrPrimaryDarker;
+}
+/**
+ * Hide element visually, but leave it available for screenreaders
+ * anything that gets toggled but needs to be accessible
+ * is a good candidate for this instead of display none
+ * $see http://open.blogs.nytimes.com/2014/06/24/improving-article-accessibility/
+ */
+.visually-hidden {
+	background: $clrWhite;
+	border: 0;
+	clip: rect(0 0 0 0);
+	color: $clrPrimaryDarker;
+	height: 1px;
+	margin: -1px;
+	overflow: hidden;
+	padding: 0;
+	position: absolute;
+	width: 1px;
+}
+
+.visually-hidden:focus {
+	border: 1px solid;
+	clip: auto;
+	display: block;
+	height: auto;
+	margin: 0;
+	overflow: visible;
+	padding: 10px;
+	position: static;
+	width: auto;
+}
+
+.select-wrap {
+	position: relative;
+}
+
+.select-wrap:after {
+	absolute: top 12px right 10px;
+	background: url("/img/arrow-down.png") no-repeat;
+	background-size: auto 15px;
+	content: '';
+	cursor: pointer;
+	display: block;
+	height: 15px;
+	pointer-events: none;
+	position: absolute;
+	right: 10px;
+	top: 12px;
+	width: 15px;
+	z-index: 2;
+}
+
+.menu-item {
+	color: $clrBase;
+	border-radius: 0;
+	font-size: $pSize;
+	padding: 10px 10px 10px 0px;
+	text-align: left;
+}
+
+.menu-item.depth-2,
+.menu-item.depth-3 {
+	font-size: $smallSize;
+}
+
+.menu-item.depth-2 {
+	padding-left: 20px;
+}
+
+.menu-item.depth-3 {
+	padding-left: 30px;
+}
+
+.menu-item.active {
+	color: $clrPrimary;
+	border-left: 4px solid $clrPrimary;
+	font-weight: bolder;
+	padding-left: 15px;
+}
+
+.menu-item a:hover {
+	background-color: $clrGrayLightest;
+}
+
+.link-external {
+	background-image: url("/img/i_external.png");
+	background-repeat: no-repeat;
+	background-position: 120% 8px;
+	background-size: 15px;
+}
+
+.line-clamp {
+	-webkit-box-orient: vertical;
+	-webkit-line-clamp: 3;
+	display: -webkit-box;
+	max-height: 80px;
+}
+
+.responsive-header {
+	font-size: 1em;
+	font-weight: 300;
+}
+
+@media (max-width: $bpTab) {
+	.responsive-header {
+		font-size: 4.5vw;
+		font-weight: bold;
+	}
+}
+
+@media (min-width: $bpTabRange1) and (max-width: $bpTabRange2) {
+	.responsive-header {
+		font-size: 22px;
+	}
+}
