+++ src/css/_syntax.scss
@@ -1,0 +1,100 @@
+/*
+github.com style (c) Vasily Polovnyov <vast@whiteants.net>
+*/
+.hljs {
+	border-radius: 3px;
+	display: inline-block;
+	font-weight: 400;
+	line-height: 1.4;
+	font-size: 14px;
+	max-width: 100%;
+	padding: 20px;
+	white-space: pre-wrap;
+	width: 100%;
+}
+
+.hljs-comment,
+.hljs-quote {
+	color: #998;
+	font-style: italic;
+}
+
+.hljs-keyword,
+.hljs-selector-tag,
+.hljs-subst {
+	color: #333;
+	font-weight: bold;
+}
+
+.hljs-number,
+.hljs-literal,
+.hljs-variable,
+.hljs-template-variable,
+.hljs-tag .hljs-attr {
+	color: #205493;
+}
+
+.hljs-string,
+.hljs-doctag {
+	color: #981b1e;
+}
+
+.hljs-title,
+.hljs-section,
+.hljs-selector-id {
+	color: #900;
+	font-weight: bold;
+}
+
+.hljs-subst {
+	font-weight: normal;
+}
+
+.hljs-type,
+.hljs-class .hljs-title {
+	color: #458;
+	font-weight: bold;
+}
+
+.hljs-tag,
+.hljs-name,
+.hljs-attribute {
+	color: #000080;
+	font-weight: normal;
+}
+
+.hljs-regexp,
+.hljs-link {
+	color: #009926;
+}
+
+.hljs-symbol,
+.hljs-bullet {
+	color: #990073;
+}
+
+.hljs-built_in,
+.hljs-builtin-name {
+	color: #0086b3;
+}
+
+.hljs-meta {
+	color: #999;
+	font-weight: bold;
+}
+
+.hljs-deletion {
+	background: #fdd;
+}
+
+.hljs-addition {
+	background: #dfd;
+}
+
+.hljs-emphasis {
+	font-style: italic;
+}
+
+.hljs-strong {
+	font-weight: bold;
+}
