+++ src/css/app.styl
@@ -8,4 +8,4 @@
 @require './_syntax'
 @require './_update'
 
-@require './components/**/*.styl'
+// @require './components/**/*.styl'
