+++ src/css/components/Nav.scss
@@ -13,7 +13,7 @@
 }
 
 .blue-nav-bar {
-	display: flex;
+	display: flex !important;
 	flex-direction: row;
 	justify-content: space-between;
 	align-items: center;
